requirements.txt
torch
torchvision
opencv-python
Pillow
scikit-learn
tqdm
facenet-pytorch
numpy 